@@ -23,27 +23,26 @@
     "e2e": "ng e2e",
     "electron": "node_modules/.bin/electron . --enable-logging --remote-debugging-port=9222",
     "start": "npm run build && npm run electron",
-    "dev": "npm run electron & ng serve"
+    "dev": "npm run electron & ng serve --aot"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.9",
-    "@angular/cdk": "^5.0.0-rc0",
-    "@angular/common": "^5.2.9",
-    "@angular/compiler": "^5.2.9",
-    "@angular/core": "^5.2.9",
-    "@angular/forms": "^5.2.9",
-    "@angular/http": "^5.2.9",
-    "@angular/material": "^5.0.0-rc0",
-    "@angular/platform-browser": "^5.2.9",
-    "@angular/platform-browser-dynamic": "^5.2.9",
-    "@angular/platform-server": "^5.2.9",
-    "@angular/router": "^5.2.9",
-    "@ngrx/store": "^5.2.0",
+    "@angular/animations": "6.0.3",
+    "@angular/cdk": "^6.2.0",
+    "@angular/common": "6.0.3",
+    "@angular/compiler": "6.0.3",
+    "@angular/core": "6.0.3",
+    "@angular/forms": "6.0.3",
+    "@angular/http": "6.0.3",
+    "@angular/material": "^6.2.0",
+    "@angular/platform-browser": "6.0.3",
+    "@angular/platform-browser-dynamic": "6.0.3",
+    "@angular/platform-server": "6.0.3",
+    "@angular/router": "6.0.3",
+    "@ngrx/store": "^6.0.1",
     "@types/cuid": "^1.3.0",
     "@types/lodash": "^4.14.85",
     "@types/pouchdb": "^6.3.2",
-    "angular2-moment": "^1.7.0",
     "core-js": "^2.4.1",
     "cuid": "^2.1.0",
     "docuri": "^4.2.2",
@@ -54,19 +53,20 @@
     "keytar": "^4.2.1",
     "lodash": "^4.17.4",
     "ngx-fs": "0.0.1",
+    "ngx-moment": "^2.0.0",
     "path": "^0.12.7",
     "photonkit": "^0.1.2",
     "pouchdb": "^6.4.3",
     "pouchdb-authentication": "^1.1.2",
-    "rxjs": "^5.5.7",
+    "rxjs": "^6.2.0",
     "split.js": "^1.3.5",
     "tinymce": "^4.7.9",
-    "zone.js": "^0.8.14"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.5.0",
-    "@angular/compiler-cli": "^5.2.9",
-    "@angular/language-service": "^5.0.1",
+    "@angular/cli": "6.0.7",
+    "@angular/compiler-cli": "6.0.3",
+    "@angular/language-service": "6.0.3",
     "@types/jasmine": "^2.8.2",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^9.6.0",
@@ -89,6 +89,7 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "^2.4.2"
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.6"
   }
 }
