@@ -2,8 +2,7 @@ import { Injectable, EventEmitter } from '@angular/core';
 import PouchDB from 'pouchdb';
 import PouchAuth from 'pouchdb-authentication';
 import { ElectronService } from 'ngx-electron';
-import { Observable } from 'rxjs/Observable';
-import { Subject } from 'rxjs/Subject';
+import { Observable ,  Subject } from 'rxjs';
 const path = require('path');
 PouchDB.plugin(PouchAuth);
 
